requirements.txt
streamlit
fastapi
redis-om
python-dotenv